requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
django-jsonfield==1.4.1
gunicorn==20.0.4
idna==2.10
lxml==4.6.2
psycopg2-binary==2.8.6
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
djangorestframework
django_filter